+++ healthtart/src/main/java/com/dev5ops/healthtart/recommended_workout_history/domain/vo/response/ResponseRegisterRecommendedWorkoutHistoryVO.java
@@ -11,7 +11,7 @@ import java.time.LocalDateTime;
 @Getter
 public class ResponseRegisterRecommendedWorkoutHistoryVO {
     private Long historyCode;
-    private int routineRatings;
+    private Double routineRatings;
     private int recommendedTime;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
